requirements.txt
aiobotocore==2.4.1
aiofiles==0.8.0
aiohttp==3.8.3
aioitertools==0.11.0
aiomysql==0.1.1
aioredis==2.0.1
aiosignal==1.3.1
aiosmtplib==2.0.0
alembic==1.9.0
amqp==5.1.1
anyio==3.6.2
asgi-csrf==0.9
asgiref==3.5.2
async-sqlalchemy==1.0.0
async-timeout==4.0.2
asyncpg==0.27.0
asyncmy==0.2.10
attrs==22.1.0
authlib==1.3.0
bandit==1.7.4
bcrypt==3.2.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
blinker==1.5
boto3==1.24.59
botocore==1.27.59
cachetools==5.2.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.4
cssselect==1.2.0
cssutils==2.6.0
clamd==1.0.2
databases==0.6.2
Deprecated==1.2.13
distlib==0.3.4
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
emails==0.6
fakeredis==2.3.0
fastapi==0.88.0
fastapi-csrf-protect==0.2.2
fastapi-login==1.8.3
fastapi-mail==0.4.1
fastapi-pagination==0.11.1
fastapi-redis-cache==0.2.5
ffmpeg-python==0.2.0
filelock==3.7.1
flower==1.2.0
frozenlist==1.3.3
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
httpx-oauth==0.7.0
humanize==4.4.0
identify==2.5.1
idna==3.4
importlib-metadata==4.12.0
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
lxml==4.9.2
mailgun3-python==0.1.4
makefun==1.13.1
Mako==1.2.4
MarkupSafe==2.1.1
minio==7.1.14
multidict==6.0.3
munch==4.0.0
mysql-connector==2.2.9
mysqlclient==2.1.1
nodeenv==1.7.0
numpy==1.26.0
opencv-python==4.8.0.76
orjson==3.8.3
orm==0.3.1
packaging==21.3
pandas==2.1.1
passlib==1.7.4
pbr==5.10.0
pillow==11.1.0
platformdirs==2.5.2
pre-commit==2.19.0
premailer==3.10.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pytest>=7.0.0
pytest-celery>=1.0.0
pytest-asyncio==0.26.0
redis==4.4.0
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
s3transfer==0.6.0
scenedetect==0.6.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
starlette==0.22.0
stevedore==4.0.0
toml==0.10.2
tornado==6.2
tqdm==4.66.1
typesystem==0.3.1
typing_extensions==4.4.0
tzdata==2022.2
tzlocal==4.2
ujson==5.6.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
vine==5.0.0
virtualenv==20.15.1
watchdog==3.0.0
watchfiles==0.18.1
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.4
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.2
zipp==3.8.0


pytest-asyncio== 0.26.0 # After installing it via pip
pytest-mock>=3.10.0